+++ aws_lambda_powertools/utilities/kafka/serialization/pydantic.py
@@ -14,9 +14,6 @@ if TYPE_CHECKING:
 
 class PydanticOutputSerializer(OutputSerializerBase):
     def serialize(self, data: dict[str, Any], output: type[T] | Callable | None = None) -> T | dict[str, Any]:
-        if output is None:
-            return data
-
         # Use TypeAdapter for better support of Union types and other complex types
         adapter: TypeAdapter = TypeAdapter(output)
         return adapter.validate_python(data)
